fix(knowledge-score): address PR review blockers
- B1: fix gofmt alignment on knowledge_score row in repository_host_tags_unit_test.go
- B2: remove dead WithBannerRepository/WithSNMPRepository wiring from manager.go
- B3: replace dead KnowledgeService with score.go/score_test.go (CalculateScore + ScoreInput only)
- B4: add backfill UPDATE to migration 016_knowledge_score.sql
- B5: add Swagger annotation to GetStatsSummary documenting knowledge score fields
- B6: add avg_knowledge_score and knowledge_score_distribution to StatsSummary in use-dashboard.ts
